requirements.txt
Flask
supabase
python-dotenv
reportlab
Werkzeug
Jinja2
gunicorn
